- Refactor settings package
