+++ app/src/main/res/layout/fragment_change_password.xml
@@ -8,7 +8,7 @@
         android:layout_height="match_parent"
         android:background="@color/app_color"
         android:orientation="vertical"
-        tools:context=".settings.changePassword.ui.ChangePasswordFragment">
+        tools:context=".settings.ui.changepassword.ChangePasswordFragment">
 
         <androidx.appcompat.widget.AppCompatImageButton
             android:id="@+id/btn_back"
