+++ app/src/main/res/layout/fragment_login_security.xml
@@ -8,7 +8,7 @@
         android:layout_height="match_parent"
         android:background="@color/app_color"
         android:orientation="vertical"
-        tools:context=".settings.loginSecurity.ui.LoginSecurityFragment">
+        tools:context=".settings.ui.loginsecurity.LoginSecurityFragment">
 
         <androidx.appcompat.widget.AppCompatImageButton
             android:id="@+id/btn_back"
